@@ -11,15 +11,15 @@ android {
         }
     }
 
-    compileSdkVersion 28
+    compileSdkVersion 31
     defaultConfig {
         applicationId "com.moh.clinicalguideline"
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 31
         versionCode 7
         versionName "1.0.6"
         signingConfig signingConfigs.config
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         vectorDrawables.useSupportLibrary = true
     }
 
@@ -42,26 +42,27 @@ android {
 }
 ext {
     roomVersion = "1.1.1"
-    daggerVer = 2.15
+    daggerVer = "2.35.1"
 }
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0-alpha3'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:design:28.0.0-alpha3'
-    implementation 'com.android.support:support-vector-drawable:28.0.0-alpha3'
-    implementation 'com.android.support:cardview-v7:28.0.0-alpha3'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
+    implementation 'com.google.android.material:material:1.5.0'
+    implementation 'androidx.vectordrawable:vectordrawable:1.1.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
     //region Install Room: SQLLite based Database persitance library
-    implementation 'com.android.support:support-v4:28.0.0-alpha3'
-    implementation 'android.arch.lifecycle:extensions:1.1.1'
-    annotationProcessor "android.arch.persistence.room:compiler:${roomVersion}"
-    implementation "android.arch.persistence.room:runtime:${roomVersion}"
-    implementation "android.arch.persistence.room:rxjava2:${roomVersion}"
-    testImplementation "android.arch.persistence.room:testing:${roomVersion}"
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
+    annotationProcessor 'androidx.room:room-compiler:2.4.2'
+    implementation 'androidx.room:room-runtime:2.4.2'
+    implementation 'androidx.room:room-rxjava2:2.4.2'
+    testImplementation 'androidx.room:room-testing:2.4.2'
     //endregion
-    implementation "android.arch.lifecycle:extensions:${roomVersion}"
-    implementation "android.arch.lifecycle:viewmodel:${roomVersion}"
-    //region Install Dagger: Dependency Injenction library
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel:2.4.1'
+    //region Install Dagger: Dependency Injection library
     implementation "com.google.dagger:dagger-android:${daggerVer}"
     implementation "com.google.dagger:dagger-android-support:${daggerVer}"
     annotationProcessor "com.google.dagger:dagger-android-processor:${daggerVer}"
@@ -75,12 +76,14 @@ dependencies {
 
     //region Install Testing Libraries
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
     //endregion
 
     //region db debugger
     debugImplementation 'com.amitshekhar.android:debug-db:1.0.4'
     //endregion
 
+    annotationProcessor 'com.android.databinding:compiler:3.2.0-alpha10'
+    implementation 'com.google.android.material:material:1.0.0'
 }
