@@ -1,12 +1,12 @@
 package com.moh.clinicalguideline.data;
 
 
-import android.arch.persistence.db.SupportSQLiteDatabase;
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.RoomDatabase;
-import android.arch.persistence.room.TypeConverters;
-import android.arch.persistence.room.migration.Migration;
-import android.support.annotation.NonNull;
+import androidx.sqlite.db.SupportSQLiteDatabase;
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
+import androidx.room.TypeConverters;
+import androidx.room.migration.Migration;
+import androidx.annotation.NonNull;
 
 import com.moh.clinicalguideline.data.dao.NodeDao;
 import com.moh.clinicalguideline.data.dao.NodeRelationDao;
