@@ -3,7 +3,7 @@ package com.moh.clinicalguideline.helper.recyclerview;
 import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.drawable.Drawable;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import com.moh.clinicalguideline.R;
 
