@@ -2,7 +2,7 @@ package com.moh.clinicalguideline.views.main;
 
 import android.content.Intent;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.moh.clinicalguideline.R;
