@@ -1,6 +1,6 @@
 package com.moh.clinicalguideline.data;
 
-import android.arch.persistence.room.TypeConverter;
+import androidx.room.TypeConverter;
 
 import java.util.Date;
 import java.util.UUID;
