@@ -1,12 +1,11 @@
 package com.moh.clinicalguideline.data.dao;
 
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.OnConflictStrategy;
-import android.arch.persistence.room.Query;
-import android.arch.persistence.room.Update;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
+import androidx.room.Query;
+import androidx.room.Update;
 
-import com.moh.clinicalguideline.data.entities.Node;
 import com.moh.clinicalguideline.data.entities.NodeRelation;
 
 import java.util.List;
