@@ -1,11 +1,10 @@
 package com.moh.clinicalguideline.data.entities;
 
-import android.arch.persistence.room.Entity;
-import android.arch.persistence.room.ForeignKey;
-import android.arch.persistence.room.Index;
-import android.arch.persistence.room.PrimaryKey;
+import androidx.room.Entity;
+import androidx.room.ForeignKey;
+import androidx.room.Index;
+import androidx.room.PrimaryKey;
 
-import java.io.NotActiveException;
 import java.util.UUID;
 
 @Entity(indices = {
