@@ -1,6 +1,6 @@
 package com.moh.clinicalguideline.helper;
 
-import android.arch.lifecycle.MutableLiveData;
+import androidx.lifecycle.MutableLiveData;
 import android.content.Context;
 
 import org.json.JSONArray;
