@@ -1,8 +1,8 @@
 package com.moh.clinicalguideline.views.algorithm;
 
 import android.annotation.SuppressLint;
-import android.arch.lifecycle.MutableLiveData;
-import android.databinding.Bindable;
+import androidx.lifecycle.MutableLiveData;
+import androidx.databinding.Bindable;
 import android.util.Log;
 import android.webkit.WebResourceError;
 import android.webkit.WebResourceRequest;
